refactor: Separate item from FileTreePane
